Refactored + added in new features in description
I have tried doing a lot of different updates to the app now. There is a way to access the camera but the way I had working intially, I have commented out as it would not add the image to the video correctly. I am working on a new way now that has not been completed just yet. 
I am working on a new way to add channels. If I get this going right, I will be able to convert my adding of videos to take a dropdown menu of all the available channels instead of just typing in manually the name of the channel.
I have a star rating system included in my system now