Final Version
This is the final version to be submitted. There are many features such as ui theme, camera and more that are set up and close to being involved, but I ran out of time to fully implament them. I hope that you'll be able to see the work that I put in to this to try to add in the features and if you feel I was close, hopefully that will be able to give me at least some partial marks.